Improve HStore lookup performance
Save the memory address of the library file with HStore functions as
well as the memory addresses of the required HStore functions, making
them members of the ArgumentData struct, now named FormatargInfoData.
Additionally, save the HStore Oid as a member of FormatargInfoData

Consequently, is_hstore() calls load_external_function() only the first time it is
called. Additionally, hstore_lookup() calls the less expensive
lookup_external_function() instead of load_external_function() and only
does so the first time it is called.